Revert "merge upstream/master"
This reverts commit 8332f5d8029f8b7a6f632249d9534b90f0b21277.

+++ gframe/client_card.h
@@ -11,7 +11,7 @@ namespace ygo {
 struct CardData {
 	unsigned int code;
 	unsigned int alias;
-	unsigned long long setcode;
+	unsigned int setcode;
 	unsigned int type;
 	unsigned int level;
 	unsigned int attribute;
@@ -22,7 +22,7 @@ struct CardData {
 struct CardDataC {
 	unsigned int code;
 	unsigned int alias;
-	unsigned long long setcode;
+	unsigned int setcode;
 	unsigned int type;
 	unsigned int level;
 	unsigned int attribute;
